refactor(lightbox): move state module to features/lightbox/state
