+++ src/view/com/lightbox/Lightbox.tsx
@@ -2,7 +2,7 @@ import {useCallback} from 'react'
 
 import {shareImageModal} from '#/lib/media/manip'
 import {useSaveImageToMediaLibrary} from '#/lib/media/save-image'
-import {useLightbox, useLightboxControls} from '#/state/lightbox'
+import {useLightbox, useLightboxControls} from '#/features/lightbox/state'
 import ImageView from './ImageViewing'
 
 export function Lightbox() {
